chore: migrate validators
