+++ plugins/module_utils/main/dhcp_controlagent.py
@@ -2,7 +2,7 @@ from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.ansibleguy.opnsense.plugins.module_utils.base.api import \
     Session
-from ansible_collections.ansibleguy.opnsense.plugins.module_utils.helper.main import \
+from ansible_collections.ansibleguy.opnsense.plugins.module_utils.helper.validate import \
     is_ip, validate_port
 from ansible_collections.ansibleguy.opnsense.plugins.module_utils.base.cls import GeneralModule
