+++ plugins/module_utils/main/frr_bfd_neighbor.py
@@ -2,7 +2,7 @@ from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.ansibleguy.opnsense.plugins.module_utils.base.api import \
     Session
-from ansible_collections.ansibleguy.opnsense.plugins.module_utils.helper.main import \
+from ansible_collections.ansibleguy.opnsense.plugins.module_utils.helper.validate import \
     is_ip_or_network
 from ansible_collections.ansibleguy.opnsense.plugins.module_utils.base.cls import BaseModule
