+++ plugins/module_utils/main/gateway.py
@@ -2,7 +2,7 @@ from ipaddress import ip_address
 
 from ansible.module_utils.basic import AnsibleModule
 
-from ansible_collections.ansibleguy.opnsense.plugins.module_utils.helper.main import \
+from ansible_collections.ansibleguy.opnsense.plugins.module_utils.helper.validate import \
     validate_int_fields, is_ip6
 from ansible_collections.ansibleguy.opnsense.plugins.module_utils.base.api import \
     Session
