+++ plugins/module_utils/main/wireguard_server.py
@@ -2,7 +2,7 @@ from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.ansibleguy.opnsense.plugins.module_utils.base.api import \
     Session
-from ansible_collections.ansibleguy.opnsense.plugins.module_utils.helper.main import \
+from ansible_collections.ansibleguy.opnsense.plugins.module_utils.helper.validate import \
     validate_int_fields, validate_str_fields, is_ip, validate_port, is_ip_or_network, \
     is_unset
 from ansible_collections.ansibleguy.opnsense.plugins.module_utils.main.wireguard_peer import Peer
